requirements.txt
-e .
langchain
openai
fastapi
pydantic
uvicorn
python-dotenv
cloudinary